Fix bug #2210, but not on Address Book
Bug report is titled 'Display name with invalid addresses not quoted
when inserted from address book', but the real problem is not the AB.

AB doesn't care about the address format or validity, the problem is
sending those addresses to the SMTP server, which, as seen in the wild¹,
might care.

Fix quotes all addresses which need quoting just before writting the
message to the queue folder.

¹http://lists.claws-mail.org/pipermail/users/2013-October/007485.html