Fix bug #3263 ‘Wrong timezone while using date_fmt function’
In addition of %z for timezone offset, the strftime allows also a %Z
specifier for timezone name and also some modifiers and field width.

All these specifier are detected now and ignored, appending the timezone
on the message. Otherwise they would be passed to the strftime call as
is, which uses the local timezone, wrong most of the time.