Added config versioning.
This allows us to change meaning of existing preferences,
without losing or mangling user data. Controlled by a common
pref "config_version", which is checked on startup after
reading all configuration. If this version is lower than
what is defined in CLAWS_CONFIG_VERSION, an upgrade is done
incrementally, from one version to the next and to the next,
so that all affected preferences are adjusted.