Fix potential crash in vcalendar's icalcomponent_vanew() on 64-bit archs.
We need to pass the terminating zero cast to void*,
because size of void* is 8 bytes, while size of a
literal 0 just 4 bytes. The arguments are retrieved
as void*, so without this void* cast, we're taking
four extra bytes belonging to whatever data is
stored in memory directly after our four-byte zero.