fix using building with --disable-deprecated
If --disable-deprecated is specified it will set GTK_CFLAGS.
Later the pkgconfig check for GTK_CFLAGS will be skipped because the
variable is already set. As a result essential CLFAGS from gtk+-2.0 will
be missing in global CFLAGS and compilation fails.

Solution is to check for the configure flag after pkgconfig.

Thanks to Olaf Hering <olaf@aepfle.de>