Handle config_version update separately for accounts.
Instead of loading entire configuration from all the
files in rcdir and doing an update at the end, we
now upgrade each "part" right after we load it.

This is to allow greater flexibility during the time
not all of the configuration is loaded yet, e.g. folders
using correct folderclass functions, based on protocol
of account owning a particular folder.

This is currently affecting clawsrc and accountrc, but
it will be trivial to extend it to other rcdir files
as needed.