Fix bug #3881: Keep read thread during update if it is selected.
When new emails arrive, the summary list is updated and the read
threads are removed if the option 'Hide read threads' is checked.

However, if you are reading the last email of a thread, then it
automatically goes to the next email. If the new selected email
happens to be the last read in a thread, then the next update
will do the same even though the user might not have seen that
message yet.

This patch fixes it to keep the thread if a message is selected.