Disallow newline character in POP3 account username or password.
The entries will have red background and preferences
will refuse to be saved, with an alertpanel telling why,
if a newline is found in either of the two entries.

This is only checked for POP3 protocol, as other protocols
encode passwords e.g. in base64, and so will not break
the communication to server by sending the extraneous
newline verbatim.