Set flags of all selected messages before updating view
If a collapsed, unread thread is marked as 'ignore thread' the first
message remains bold in the message list. That happens because
summary_ignore_thread_func calls summary_set_row_marks, which finds
unread children. As a result bold_style is selected.

To fix this it is required to first update the message flags of all
selected messages, before updating the view of all selected messages.

Signed-off-by: Olaf Hering <olaf@aepfle.de>