Fix bug 3793: segfault when autocompletion asks for master passphrase
This is done by checking early if any ldap server is password-protected,
and asking for master passphrase when compose window appears.  If user
cancels the dialog (does not enter the passphrase), we temporarily
disable the password-protected LDAP servers, just for that particular
compose window. We reenable them in compose_destroy().