fix bug 3973, ''select all' in summaryview does not automatically focus the summaryview'
The problem is apparent when 'default selection on entering a folder' is
set to 'none' and the first user action on entering the folder is
'select all'. This visually selected all the messages but summaryview
did not have the focus, so any actions performed on that list would
fail.

Removed summary_select_all(), which didn't quite do as advertised and
instead use gtk_cmclist_select_all() which selects and focuses.