fix bug where creating a new folder wth inheriting properties doesn't write processinf rules
after copying folder with 'inherit properties from parent folder' set,
the processing rules are (initially) present in the processing rules
dialogue, but are not actually written. So on a restart they are 'lost'.

This is probably the actual bug 4121, ' Moving a subfolder in another
folder erases  its processing rules', rather than the description given.