Implement an LRU pixbuf cache for LiteHTML viewer
Replace the list-based pixbuf cache with a map-based system.  Messages with
duplicate image urls were causing multiple downloads and cache entries for
the same url.  The map-based cache enforces uniqueness based on the url.

Also add a timestamp to the cache entries so that they can be pruned based on
the last requested time, rather than the order they were first requested.