Handle more systems which lack support for the scanf m modifier
The 'm' modifier for scanf is an extension to ISO C. It was mentioned in
POSIX 1003.1 2008 (Issue 7). It is difficult to detect at compile time
if a libc has implemented support for this modifier. Extend the existing
list of systems which lack support for it with NetBSD and Dragonfly BSD.

Signed-off-by: Olaf Hering <olaf@aepfle.de>