postpone auto-updates if Retry-After: given
Sometimes when a feed returns an HTTP error code due to having received
too many requests in too short a period of time (especially 403 or 429),
they specify a safe interval after which the request could be retried
using the Retry-After: HTTP header. Parse the value and skip automatic
updates if it's set to a date in the future.

When a manual update succeeds, clear the retry_after value, making it
possible to unstuck automatic updates in case a Retry-After: value was
accidentally set too far in the future.

patch by Ivan Krylov